+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix missing translations in the user interface of the plugin
 * Improve the panel about server information
 * Improve metadata in the CFG file about the status of the current Lizmap Web Client version which is used
+* Improve a little the UX about the initial scales
 
 ## 3.9.0 - 2022-07-25